refactored drive commands
