+++ src/main/java/frc/robot/commands/SwerveDriveCommands.java
@@ -1,25 +1,18 @@
 package frc.robot.commands;
 
-import java.text.DecimalFormat;
-import java.text.NumberFormat;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.math.controller.ProfiledPIDController;
-import edu.wpi.first.math.filter.SlewRateLimiter;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Transform2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.math.trajectory.TrapezoidProfile;
-import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
-import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import frc.robot.subsystems.SwerveDriveSubsystem;
@@ -33,10 +26,6 @@ public class SwerveDriveCommands {
     private static final double ANGLE_KD = 0.4;
     private static final double ANGLE_MAX_VELOCITY = 8.0;
     private static final double ANGLE_MAX_ACCELERATION = 20.0;
-    private static final double FF_START_DELAY = 2.0; // Secs
-    private static final double FF_RAMP_RATE = 0.1; // Volts/Sec
-    private static final double WHEEL_RADIUS_MAX_VELOCITY = 0.25; // Rad/Sec
-    private static final double WHEEL_RADIUS_RAMP_RATE = 0.05; // Rad/Sec^2
 
     /**
      * 
@@ -142,149 +131,5 @@ public class SwerveDriveCommands {
 
                 // Reset PID controller when command starts
                 .beforeStarting(() -> angleController.reset(swerveDriveSubsystem.getRotation().getRadians()));
-    }
-
-    /**
-     * Measures the velocity feedforward constants for the drive motors.
-     *
-     * <p>
-     * This command should only be used in voltage control mode.
-     */
-    public static Command feedforwardCharacterization(SwerveDriveSubsystem swerveDriveSubsystem) {
-        List<Double> velocitySamples = new LinkedList<>();
-        List<Double> voltageSamples = new LinkedList<>();
-        Timer timer = new Timer();
-
-        return Commands.sequence(
-                // Reset data
-                Commands.runOnce(() -> {
-                    velocitySamples.clear();
-                    voltageSamples.clear();
-                }),
-
-                // Allow modules to orient
-                Commands.run(() -> {
-                    swerveDriveSubsystem.runCharacterization(0.0);
-                },
-                        swerveDriveSubsystem)
-                        .withTimeout(FF_START_DELAY),
-
-                // Start timer
-                Commands.runOnce(timer::restart),
-
-                // Accelerate and gather data
-                Commands.run(
-                        () -> {
-                            double voltage = timer.get() * FF_RAMP_RATE;
-                            swerveDriveSubsystem.runCharacterization(voltage);
-                            velocitySamples.add(swerveDriveSubsystem.getFFCharacterizationVelocity());
-                            voltageSamples.add(voltage);
-                        },
-                        swerveDriveSubsystem)
-
-                        // When cancelled, calculate and print results
-                        .finallyDo(
-                                () -> {
-                                    int n = velocitySamples.size();
-                                    double sumX = 0.0;
-                                    double sumY = 0.0;
-                                    double sumXY = 0.0;
-                                    double sumX2 = 0.0;
-                                    for (int i = 0; i < n; i++) {
-                                        sumX += velocitySamples.get(i);
-                                        sumY += voltageSamples.get(i);
-                                        sumXY += velocitySamples.get(i) * voltageSamples.get(i);
-                                        sumX2 += velocitySamples.get(i) * velocitySamples.get(i);
-                                    }
-                                    double kS = (sumY * sumX2 - sumX * sumXY) / (n * sumX2 - sumX * sumX);
-                                    double kV = (n * sumXY - sumX * sumY) / (n * sumX2 - sumX * sumX);
-
-                                    NumberFormat formatter = new DecimalFormat("#0.00000");
-                                    System.out.println("********** Drive FF Characterization Results **********");
-                                    System.out.println("\tkS: " + formatter.format(kS));
-                                    System.out.println("\tkV: " + formatter.format(kV));
-                                }));
-    }
-
-    /**
-     * Measures the robot's wheel radius by spinning in a circle.
-     */
-    public static Command wheelRadiusCharacterization(SwerveDriveSubsystem swerveDriveSubsystem) {
-        SlewRateLimiter limiter = new SlewRateLimiter(WHEEL_RADIUS_RAMP_RATE);
-        WheelRadiusCharacterizationState state = new WheelRadiusCharacterizationState();
-
-        return Commands.parallel(
-                // Drive control sequence
-                Commands.sequence(
-                        // Reset acceleration limiter
-                        Commands.runOnce(
-                                () -> {
-                                    limiter.reset(0.0);
-                                }),
-
-                        // Turn in place, accelerating up to full speed
-                        Commands.run(
-                                () -> {
-                                    double speed = limiter.calculate(WHEEL_RADIUS_MAX_VELOCITY);
-                                    swerveDriveSubsystem.runVelocity(new ChassisSpeeds(0.0, 0.0, speed));
-                                },
-                                swerveDriveSubsystem)),
-
-                // Measurement sequence
-                Commands.sequence(
-                        // Wait for modules to fully orient before starting measurement
-                        Commands.waitSeconds(1.0),
-
-                        // Record starting measurement
-                        Commands.runOnce(
-                                () -> {
-                                    state.positions = swerveDriveSubsystem.getWheelRadiusCharacterizationPositions();
-                                    state.lastAngle = swerveDriveSubsystem.getRotation();
-                                    state.gyroDelta = 0.0;
-                                }),
-
-                        // Update gyro delta
-                        Commands.run(
-                                () -> {
-                                    var rotation = swerveDriveSubsystem.getRotation();
-                                    state.gyroDelta += Math.abs(rotation.minus(state.lastAngle).getRadians());
-                                    state.lastAngle = rotation;
-                                })
-
-                                // When cancelled, calculate and print results
-                                .finallyDo(
-                                        () -> {
-                                            double[] positions = swerveDriveSubsystem
-                                                    .getWheelRadiusCharacterizationPositions();
-                                            double wheelDelta = 0.0;
-                                            for (int i = 0; i < 4; i++) {
-                                                wheelDelta += Math.abs(positions[i] - state.positions[i]) / 4.0;
-                                            }
-                                            double wheelRadius = (state.gyroDelta
-                                                    * SwerveDriveSubsystem.DRIVE_BASE_RADIUS) / wheelDelta;
-
-                                            NumberFormat formatter = new DecimalFormat("#0.000");
-                                            System.out.println(
-                                                    "********** Wheel Radius Characterization Results **********");
-                                            System.out.println(
-                                                    "\tWheel Delta: " + formatter.format(wheelDelta) + " radians");
-                                            System.out.println(
-                                                    "\tGyro Delta: " + formatter.format(state.gyroDelta) + " radians");
-                                            System.out.println(
-                                                    "\tWheel Radius: "
-                                                            + formatter.format(wheelRadius)
-                                                            + " meters, "
-                                                            + formatter.format(Units.metersToInches(wheelRadius))
-                                                            + " inches");
-                                        })));
-    }
-
-    /**
-     * 
-     */
-    private static class WheelRadiusCharacterizationState {
-        double[] positions = new double[4];
-        Rotation2d lastAngle = new Rotation2d();
-        double gyroDelta = 0.0;
     }
 }
