+++ src/main/java/frc/robot/commands/WheelRadiusCharacterizationCommand.java
@@ -1,0 +1,102 @@
+package frc.robot.commands;
+
+import java.text.DecimalFormat;
+import java.text.NumberFormat;
+
+import edu.wpi.first.math.filter.SlewRateLimiter;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.kinematics.ChassisSpeeds;
+import edu.wpi.first.math.util.Units;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
+import frc.robot.subsystems.SwerveDriveSubsystem;
+
+/**
+ * Measures the robot's wheel radius by spinning in a circle.
+ */
+public class WheelRadiusCharacterizationCommand {
+    private static final double WHEEL_RADIUS_MAX_VELOCITY = 0.25; // Rad/Sec
+    private static final double WHEEL_RADIUS_RAMP_RATE = 0.05; // Rad/Sec^2
+
+    /**
+     * 
+     */
+    private static class WheelRadiusCharacterizationState {
+        double[] positions = new double[4];
+        Rotation2d lastAngle = new Rotation2d();
+        double gyroDelta = 0.0;
+    }
+
+    /**
+     * 
+     */
+    public static Command get(SwerveDriveSubsystem swerveDriveSubsystem) {
+        SlewRateLimiter limiter = new SlewRateLimiter(WHEEL_RADIUS_RAMP_RATE);
+        WheelRadiusCharacterizationState state = new WheelRadiusCharacterizationState();
+
+        return Commands.parallel(
+                // Drive control sequence
+                Commands.sequence(
+                        // Reset acceleration limiter
+                        Commands.runOnce(
+                                () -> {
+                                    limiter.reset(0.0);
+                                }),
+
+                        // Turn in place, accelerating up to full speed
+                        Commands.run(
+                                () -> {
+                                    double speed = limiter.calculate(WHEEL_RADIUS_MAX_VELOCITY);
+                                    swerveDriveSubsystem.runVelocity(new ChassisSpeeds(0.0, 0.0, speed));
+                                },
+                                swerveDriveSubsystem)),
+
+                // Measurement sequence
+                Commands.sequence(
+                        // Wait for modules to fully orient before starting measurement
+                        Commands.waitSeconds(1.0),
+
+                        // Record starting measurement
+                        Commands.runOnce(
+                                () -> {
+                                    state.positions = swerveDriveSubsystem.getWheelRadiusCharacterizationPositions();
+                                    state.lastAngle = swerveDriveSubsystem.getRotation();
+                                    state.gyroDelta = 0.0;
+                                }),
+
+                        // Update gyro delta
+                        Commands.run(
+                                () -> {
+                                    var rotation = swerveDriveSubsystem.getRotation();
+                                    state.gyroDelta += Math.abs(rotation.minus(state.lastAngle).getRadians());
+                                    state.lastAngle = rotation;
+                                })
+
+                                // When cancelled, calculate and print results
+                                .finallyDo(
+                                        () -> {
+                                            double[] positions = swerveDriveSubsystem
+                                                    .getWheelRadiusCharacterizationPositions();
+                                            double wheelDelta = 0.0;
+                                            for (int i = 0; i < 4; i++) {
+                                                wheelDelta += Math.abs(positions[i] - state.positions[i]) / 4.0;
+                                            }
+                                            double wheelRadius = (state.gyroDelta
+                                                    * SwerveDriveSubsystem.DRIVE_BASE_RADIUS) / wheelDelta;
+
+                                            NumberFormat formatter = new DecimalFormat("#0.000");
+                                            System.out.println(
+                                                    "********** Wheel Radius Characterization Results **********");
+                                            System.out.println(
+                                                    "\tWheel Delta: " + formatter.format(wheelDelta) + " radians");
+                                            System.out.println(
+                                                    "\tGyro Delta: " + formatter.format(state.gyroDelta) + " radians");
+                                            System.out.println(
+                                                    "\tWheel Radius: "
+                                                            + formatter.format(wheelRadius)
+                                                            + " meters, "
+                                                            + formatter.format(Units.metersToInches(wheelRadius))
+                                                            + " inches");
+                                        })));
+    }
+}
